requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
jinja2>=3.1.4
python-multipart>=0.0.12
aiofiles>=24.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1.0
reportlab>=4.0.0
fastapi-mail>=1.4.1
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
